backend/requirements.txt
# Django
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3

# 数据库
mysqlclient==2.2.0
redis==4.6.0

# 工具
python-dotenv==1.0.0
PyJWT==2.8.0
drf-yasg==1.21.7

# 开发工具
black==23.7.0
flake8==6.1.0
pytest==7.4.2
pytest-django==4.5.2